fixed known issues. need further testing. norminetting
