+++ src/parsing/constructor_3.c
@@ -1,0 +1,89 @@
+#include "ft_parsing.h"
+
+// function to add a t_cmd to the t_cmd linked list
+void	add_last_cmd(t_cmd **head, t_cmd *new_cmd)
+{
+	t_cmd	*tmp;
+
+	tmp = *head;
+	if (!tmp)
+		*head = new_cmd;
+	else
+	{
+		while (tmp->next)
+			tmp = tmp->next;
+		tmp->next = new_cmd;
+		new_cmd->prev = tmp;
+	}
+}
+
+char	**copy_array(char **array)
+{
+	int		i;
+	char	**new_array;
+
+	i = 0;
+	while (array[i])
+		i++;
+	new_array = ft_calloc(sizeof(char *), i + 1);
+	i = 0;
+	while (array[i])
+	{
+		new_array[i] = ft_strdup(array[i]);
+		i++;
+	}
+	new_array[i] = NULL;
+	return (new_array);
+}
+
+void	check_builtin(t_cmd *cmd)
+{
+	if (!cmd)
+		return ;
+	if (ft_strcmp(cmd->name, "env") == 0)
+		cmd->is_builtin = true;
+	if (ft_strcmp(cmd->name, "exit") == 0)
+		cmd->is_builtin = true;
+	if (ft_strcmp(cmd->name, "export") == 0)
+		cmd->is_builtin = true;
+	if (ft_strcmp(cmd->name, "unset") == 0)
+		cmd->is_builtin = true;
+	if (ft_strcmp(cmd->name, "pwd") == 0)
+		cmd->is_builtin = true;
+	if (ft_strcmp(cmd->name, "cd") == 0)
+		cmd->is_builtin = true;
+	if (ft_strcmp(cmd->name, "echo") == 0)
+		cmd->is_builtin = true;
+	if (ft_strcmp(cmd->name, "$?") == 0)
+		cmd->is_builtin = true;
+}
+
+void	build_cmd(t_cmd *cmd, char **args, t_shell *shell)
+{
+	if (is_command_path(cmd->name))
+		return (assemble_cmd_path(cmd, args));
+	else
+		return (assemble_cmd(cmd, args, shell));
+}
+
+void	build_builtin(t_cmd *cmd, char **args)
+{
+	cmd->is_valid = true;
+	if (ft_strcmp(cmd->name, "cd") == 0)
+		cmd->builtin_func = cd_shell;
+	if (ft_strcmp(cmd->name, "echo") == 0)
+		cmd->builtin_func = echo_shell;
+	if (ft_strcmp(cmd->name, "pwd") == 0)
+		cmd->builtin_func = pwd_shell;
+	if (ft_strcmp(cmd->name, "export") == 0)
+		cmd->builtin_func = export_builtin;
+	if (ft_strcmp(cmd->name, "unset") == 0)
+		cmd->builtin_func = unset_vars;
+	if (ft_strcmp(cmd->name, "exit") == 0)
+		cmd->builtin_func = exit_shell;
+	if (ft_strcmp(cmd->name, "env") == 0)
+		cmd->builtin_func = env;
+	if (ft_strcmp(cmd->name, "$?") == 0)
+		cmd->builtin_func = exit_code_printer;
+	cmd->args = copy_array(args + 1);
+}
